+++ src/test/java/io/github/cpjust/testng_annotations/listeners/EnvListenerBaseTest.java
@@ -13,7 +13,10 @@ import java.util.Properties;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.when;
 
 // NOTE: These tests need to be unit tests and use Mockito because they throw exceptions that would cause TestNG to