+++ src/test/java/io/github/cpjust/testng_annotations/listeners/IncludeOnEnvListenerTest.java
@@ -10,7 +10,8 @@ import org.testng.IMethodInstance;
 
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(MockitoExtension.class)
 class IncludeOnEnvListenerTest extends BaseTestEnvListener {