+++ CHANGELOG.md
@@ -6,11 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
-## [1.0.1] - 2025-08-24
+## [1.1.0] - 2025-09-20
+
+### Added
+- `@ValueSource` annotation to provide a source for test values, just like JUnit 5's `@ValueSource`.
+- `ValueSourceListener` to handle the `@ValueSource` annotation and provide test values for parameterized tests.
+
+## [1.0.1] - 2025-09-19
 ### Added
 - Support for CSV strings in `@IncludeOnEnv` and `@ExcludeOnEnv` annotations by a new `delimiter` attribute.
+- A test that uses `@Test` priority to ensure that the `@IncludeOnEnv` and `@ExcludeOnEnv` work correctly with test priorities.
 
-## [1.0.0] - 2025-07-29
+## [1.0.0] - 2025-08-05
 Initial version.
 
 Contains the following features: